requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
Authlib==1.3.1
azure-ai-formrecognizer==3.3.3
azure-common==1.1.28
azure-core==1.33.0
bcrypt==3.2.0
boto3==1.38.8
botocore==1.38.8
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
elastic-transport==8.17.1
elasticsearch==8.17.2
email_validator==2.2.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi==0.115.12
filelock==3.18.0
frozenlist==1.5.0
fsspec==2025.3.2
google-api-core==2.24.2
google-auth==2.39.0
google-cloud-documentai==3.3.0
google-genai==1.11.0
googleapis-common-protos==1.70.0
google-generativeai==0.8.4
PyMySQL==1.1.1
grpcio==1.71.0
grpcio-health-checking==1.71.0
grpcio-status==1.71.0
grpcio-tools==1.71.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
importlib_metadata==8.6.1
isodate==0.7.2
Jinja2==3.1.6
jiter==0.9.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
litellm==1.66.1
MarkupSafe==3.0.2
msrest==0.7.1
multidict==6.4.3
numpy==2.2.4
oauthlib==3.2.2
openai==1.74.0
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
passlib==1.7.4
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
pyasn1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.2
pydantic-settings==2.9.1
pydantic_core==2.33.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.24.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
tiktoken==0.9.0
tokenizers==0.21.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
validators==0.34.0
weaviate-client==4.13.2
websockets==15.0.1
yarl==1.19.0
zipp==3.21.0
vanna==0.7.9
google-cloud-aiplatform==1.91.0
